minor dtype-related changes

+++ src/peft/tuners/sft/layer.py
@@ -272,7 +272,7 @@ class Linear(nn.Linear, BaseTunerLayer):
             #    result = result2
             if self.hook is None:
                 #values = F.dropout(sft.values, p=0.05, training=self.training)
-                result = linear_sd_cpp.apply(x, self.weight, sft.values, sft.indices, bias=self.bias)
+                result = linear_sd_cpp.apply(x.to(self.weight.dtype), self.weight, sft.values, sft.indices, bias=self.bias)
             else:
                 merged_weight = sft(self.weight)
                 if merged_weight.requires_grad:
